+++ instapy/instapy.py
@@ -1,6 +1,7 @@
 """OS Modules environ method to get the setup vars from the Environment"""
 # import built-in & third-party modules
 import time
+from datetime import datetime, timedelta
 from math import ceil
 import random
 from sys import platform
@@ -47,6 +48,8 @@ from .util import dump_record_activity
 from .util import truncate_float
 from .util import save_account_progress
 from .util import parse_cli_args
+from .util import get_cord_location
+from .util import get_bounding_box
 from .unfollow_util import get_given_user_followers
 from .unfollow_util import get_given_user_following
 from .unfollow_util import unfollow
@@ -72,6 +75,9 @@ from .browser import set_selenium_local_session
 from .browser import close_browser
 from .file_manager import get_workspace
 from .file_manager import get_logfolder
+from .pods_util import get_recent_posts_from_pods
+from .pods_util import share_my_post_with_pods
+from .pods_util import share_with_pods_restriction
 
 # import exceptions
 from selenium.common.exceptions import NoSuchElementException
@@ -175,6 +181,7 @@ class InstaPy:
         self.already_Visited = 0
 
         self.follow_times = 1
+        self.share_times = 1
         self.do_follow = False
         self.follow_percentage = 0
         self.dont_include = set()
@@ -184,6 +191,7 @@ class InstaPy:
         self.do_like = False
         self.like_percentage = 0
         self.smart_hashtags = []
+        self.smart_location_hashtags = []
 
         self.dont_like = ['sex', 'nsfw']
         self.mandatory_words = []
@@ -253,6 +261,8 @@ class InstaPy:
                                      "unfollows": 0},
                       "limit": {"likes": 7, "comments": 3, "follows": 5,
                                 "unfollows": 4}}
+
+        self.allowed_pod_topics = ['general', 'beauty', 'food', 'travel', 'sports', 'entertainment']
 
         # stores the features' name which are being used by other features
         self.internal_usage = {}
@@ -646,6 +656,47 @@ class InstaPy:
 
         # delete duplicated tags
         self.smart_hashtags = list(set(self.smart_hashtags))
+        return self
+
+    def set_smart_location_hashtags(self,
+                                    locations,
+                                    radius=10,
+                                    limit=3,
+                                    log_tags=True):
+        """Generate smart hashtags based on https://displaypurposes.com/map"""
+        if locations is None:
+            self.logger.error('set_smart_location_hashtags is misconfigured')
+            return self
+
+        for location in locations:
+            lat, lon = get_cord_location(self.browser, location)
+
+            bbox = get_bounding_box(lat,
+                                    lon,
+                                    logger=self.logger,
+                                    half_side_in_miles=radius)
+            bbox_url = '{},{},{},{}&zoom={}'.format(bbox['lon_min'], bbox['lat_min'], bbox['lon_max'],
+                                                    bbox['lat_max'], radius)
+            url = 'https://query.displaypurposes.com/local/?bbox={}'.format(bbox_url)
+
+            req = requests.get(url)
+            data = json.loads(req.text)
+            if int(data['count']) == 0:
+                self.logger.warning(u'Too few results for {} location'.format(location))
+                continue
+
+            count = limit if limit < data['count'] else data['count']
+            i = 0
+            tags = []
+            while i < count:
+                self.smart_location_hashtags.append(data['tags'][i]['tag'])
+                i += 1
+
+        self.smart_location_hashtags = list(set(self.smart_location_hashtags))
+
+        if log_tags is True:
+            self.logger.info(u'[smart location hashtag generated: {}]\n'.format(self.smart_location_hashtags))
+
         return self
 
     def set_mandatory_language(self, enabled=False, character_set='LATIN'):
@@ -1625,6 +1676,7 @@ class InstaPy:
                      amount=50,
                      skip_top_posts=True,
                      use_smart_hashtags=False,
+                     use_smart_location_hashtags=False,
                      interact=False,
                      randomize=False,
                      media=None):
@@ -1641,8 +1693,11 @@ class InstaPy:
 
         # if smart hashtag is enabled
         if use_smart_hashtags is True and self.smart_hashtags is not []:
-            print('Using smart hashtags')
+            self.logger.info('Using smart hashtags')
             tags = self.smart_hashtags
+        elif use_smart_location_hashtags is True and self.smart_location_hashtags is not []:
+            self.logger.info('Using smart location hashtags')
+            tags = self.smart_location_hashtags
 
         # deletes white spaces in tags
         tags = [tag.strip() for tag in tags]
@@ -4008,11 +4063,11 @@ class InstaPy:
 
         return mutual_following
 
-    def end(self):
+    def end(self, threaded_session=False):
         """Closes the current session"""
 
         Settings.InstaPy_is_running = False
-        close_browser(self.browser, False, self.logger)
+        close_browser(self.browser, threaded_session, self.logger)
 
         with interruption_handler():
             # close virtual display
@@ -4151,6 +4206,7 @@ class InstaPy:
                        amount=50,
                        skip_top_posts=True,
                        use_smart_hashtags=False,
+                       use_smart_location_hashtags=False,
                        randomize=False,
                        media=None,
                        interact=False):
@@ -4163,8 +4219,11 @@ class InstaPy:
 
         # if smart hashtag is enabled
         if use_smart_hashtags is True and self.smart_hashtags is not []:
-            print('Using smart hashtags')
+            self.logger.info('Using smart hashtags')
             tags = self.smart_hashtags
+        elif use_smart_location_hashtags is True and self.smart_location_hashtags is not []:
+            self.logger.info('Using smart location hashtags')
+            tags = self.smart_location_hashtags
 
         # deletes white spaces in tags
         tags = [tag.strip() for tag in tags]
@@ -5146,5 +5205,120 @@ class InstaPy:
                     break
 
         self.logger.info("Accepted {} follow requests".format(accepted))
+
+        return self
+
+    def join_pods(self, topic='general'):
+        """ Join pods """
+        if topic not in self.allowed_pod_topics:
+            self.logger.error('You have entered an invalid topic for pods, allowed topics are : {}. Exiting...'.format(self.allowed_pod_topics))
+            return self
+
+        if self.comments is not None and len(self.comments) < 10:
+            self.logger.error('You have too few comments, please set at least 10 distinct comments to avoid looking suspicious.')
+            return self
+
+        user_link = 'https://www.instagram.com/{}/'.format(self.username)
+        web_address_navigator(self.browser, user_link)
+        try:
+            pod_post_ids = get_recent_posts_from_pods(topic, self.logger)
+            sleep(2)
+            post_link_elems = self.browser.find_elements_by_xpath("//a[contains(@href, '/p/')]")
+            post_links = []
+
+            for post_link_elem in post_link_elems:
+                try:
+                    post_link = post_link_elem.get_attribute('href')
+                    post_links.append(post_link)
+                except Exception as e:
+                    self.logger.error('Can not get href for {} - {}'.format(post_link, e))
+                    continue
+
+            post_links = list(set(post_links))
+            my_recent_post_ids = []
+            for post_link in post_links:
+                web_address_navigator(self.browser, post_link)
+                sleep(2)
+                time_element = self.browser.find_element_by_xpath("//div/a/time")
+                post_datetime_str = time_element.get_attribute('datetime')
+                post_datetime = datetime.strptime(post_datetime_str, "%Y-%m-%dT%H:%M:%S.%fZ")
+                postid = post_link.split('/')[4]
+                self.logger.info("Post: {}, Instaposted at: {}".format(postid, post_datetime))
+                share_restricted = share_with_pods_restriction("read", postid,
+                                        self.share_times,
+                                        self.logger)
+                if datetime.now() - post_datetime < timedelta(hours=12, minutes=30) and not share_restricted:
+                    my_recent_post_ids.append(postid)
+                    if share_my_post_with_pods(postid, topic, self.logger):
+                        share_with_pods_restriction("write", postid, None, self.logger)
+
+            if len(my_recent_post_ids) > 0:
+                self.logger.info("I have recent post(s), so I will now help pod members actively.")
+                nposts = 200
+            else:
+                self.logger.info("I don't have any recent post(s), so I will just help a few pod posts and move on.")
+                nposts = 40
+
+            if len(pod_post_ids) <= nposts:
+                pod_post_ids = pod_post_ids
+            else:
+                pod_post_ids = random.sample(pod_post_ids, nposts)
+
+            for pod_post_id in pod_post_ids:
+                post_link = "https://www.instagram.com/p/{}".format(pod_post_id)
+                web_address_navigator(self.browser, post_link)
+
+                inappropriate, user_name, is_video, reason, scope = (
+                    check_link(self.browser,
+                                post_link,
+                                self.dont_like,
+                                self.mandatory_words,
+                                self.mandatory_language,
+                                self.is_mandatory_character,
+                                self.mandatory_character,
+                                self.check_character_set,
+                                self.ignore_if_contains,
+                                self.logger))
+
+                if user_name != self.username:
+                    follow_state, msg = follow_user(self.browser,
+                                                    "post",
+                                                    self.username,
+                                                    user_name,
+                                                    None,
+                                                    self.blacklist,
+                                                    self.logger,
+                                                    self.logfolder)
+
+                    self.dont_include.add(user_name)
+
+                if not inappropriate and user_name != self.username:
+                    pods_like_percent = max(80, min(100, self.like_percentage))
+                    pods_comment_percentage = max(80, min(100, self.comment_percentage))
+                    liking = (random.randint(0, 100) <= pods_like_percent)
+                    commenting = (random.randint(0, 100) <= pods_comment_percentage)
+
+                    if liking:
+                        like_state, msg = like_image(self.browser,
+                                                        user_name,
+                                                        self.blacklist,
+                                                        self.logger,
+                                                        self.logfolder)
+
+                    if commenting:
+                        comments = self.fetch_smart_comments(
+                                                        is_video,
+                                                        temp_comments=[])
+
+                        comment_state, msg = comment_image(
+                                                        self.browser,
+                                                        user_name,
+                                                        comments,
+                                                        self.blacklist,
+                                                        self.logger,
+                                                        self.logfolder)
+
+        except Exception as err:
+            self.logger.error(err)
 
         return self